Remove pgindent and use clang-format instead
This PR changes travis to use clang-format instead of
pgindent. It also changes travis to use the clang docker
image for the clang-format test as well as the license
check and prefix checks. This speeds up those tests and
paves the way for us to use more clang features for tests
in the future.